fix: clean up and shared markdown reader
