+++ ui/src/components/chat/Message.tsx
@@ -1,46 +1,11 @@
-import type { FC, ReactNode } from "react";
-import React, { memo } from "react";
-import ReactMarkdown from "react-markdown";
-import remarkBreaks from "remark-breaks";
-import remarkGfm from "remark-gfm";
+import Markdown from "@components/layout/Markdown";
+import type { FC } from "react";
+import { memo } from "react";
 
 interface MessageProps {
   type: string;
   text: string;
 }
-
-interface CodeProps extends React.HTMLAttributes<HTMLElement> {
-  inline?: boolean;
-  className?: string;
-  children?: ReactNode;
-}
-
-const CodeBlock: FC<CodeProps> = memo(
-  ({ inline, className, children, ...props }) => {
-    const content = children?.toString() || "";
-    if (inline) {
-      return (
-        <code className="bg-gray-100 p-1 rounded" {...props}>
-          {content}
-        </code>
-      );
-    }
-    if (className?.includes("language-markdown")) {
-      return (
-        <div className="prose max-w-full bg-gray-50 p-2 rounded overflow-x-auto">
-          <ReactMarkdown remarkPlugins={[remarkGfm, remarkBreaks]}>
-            {content}
-          </ReactMarkdown>
-        </div>
-      );
-    }
-    return (
-      <pre className="bg-gray-100 p-2 rounded overflow-x-auto">
-        <code className={className}>{content}</code>
-      </pre>
-    );
-  }
-);
 
 const Message: FC<MessageProps> = memo(({ type, text }) => {
   const isUser = type === "user";
@@ -64,30 +29,7 @@ const Message: FC<MessageProps> = memo(({ type, text }) => {
         alt={isUser ? "User Avatar" : "Bot Avatar"}
       />
       <div className={bubbleClass}>
-        <ReactMarkdown
-          remarkPlugins={[remarkGfm, remarkBreaks]}
-          components={{
-            p: ({ node, children, ...props }) => {
-              // Combine text nodes for analysis.
-              const textContent = React.Children.toArray(children)
-                .filter((child) => typeof child === "string")
-                .join("");
-              // Apply extra margin only if there's at least one newline character.
-              const marginClass = textContent.includes("\n") ? "mb-6" : "mb-2";
-              return (
-                <p
-                  className={`whitespace-pre-wrap break-words ${marginClass}`}
-                  {...props}
-                >
-                  {children}
-                </p>
-              );
-            },
-            code: CodeBlock,
-          }}
-        >
-          {text}
-        </ReactMarkdown>
+        <Markdown text={text} />
       </div>
     </div>
   );
